+++ src/engine/layouts/stackingcommon.ts
@@ -1,0 +1,10 @@
+import { Window } from "../engine";
+
+export class WindowBox {
+    window: Window;
+    size: number = 1;
+
+    constructor(window: Window) {
+        this.window = window;
+    }
+}